[advanced] 또 많은 .............. 함수 분리
