+++ packages/assignment-7/src/hooks/useEvents.ts
@@ -1,5 +1,5 @@
 import { useToast } from '@chakra-ui/react';
-import { useRef, useState } from 'react';
+import { ChangeEvent, useRef, useState } from 'react';
 
 export type RepeatType = 'none' | 'daily' | 'weekly' | 'monthly' | 'yearly';
 
@@ -44,7 +44,19 @@ interface UseEventsReturn {
   fetchEvents: () => Promise<void>;
   addOrUpdateEvent: () => Promise<void>;
   saveEvent: (eventData: Event) => Promise<void>;
+  editEvent: (event: Event) => void;
+  deleteEvent: (id: number) => Promise<void>;
   validateTime: (start: string, end: string) => void;
+  checkUpcomingEvents: () => Promise<void>;
+  handleStartTimeChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  handleEndTimeChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  getDaysInMonth: (year: number, month: number) => number;
+  getWeekDates: (date: Date) => Date[];
+  navigate: (direction: 'prev' | 'next') => void;
+  searchEvents: (term: string) => Event[];
+  formatWeek: (date: Date) => string;
+  formatMonth: (date: Date) => string;
+  filteredEvents: Event[];
   events: Event[];
   setEvents: React.Dispatch<React.SetStateAction<Event[]>>;
   title: string;
@@ -167,6 +179,22 @@ function useEvents(): UseEventsReturn {
     }
   };
 
+  const editEvent = (event: Event) => {
+    setEditingEvent(event);
+    setTitle(event.title);
+    setDate(event.date);
+    setStartTime(event.startTime);
+    setEndTime(event.endTime);
+    setDescription(event.description);
+    setLocation(event.location);
+    setCategory(event.category);
+    setIsRepeating(event.repeat.type !== 'none');
+    setRepeatType(event.repeat.type);
+    setRepeatInterval(event.repeat.interval);
+    setRepeatEndDate(event.repeat.endDate || '');
+    setNotificationTime(event.notificationTime);
+  };
+
   const resetForm = () => {
     setTitle('');
     setDate('');
@@ -227,6 +255,46 @@ function useEvents(): UseEventsReturn {
         isClosable: true,
       });
     }
+  };
+
+  const deleteEvent = async (id: number) => {
+    try {
+      const response = await fetch(`/api/events/${id}`, {
+        method: 'DELETE',
+      });
+
+      if (!response.ok) {
+        throw new Error('Failed to delete event');
+      }
+
+      await fetchEvents(); // 이벤트 목록 새로고침
+      toast({
+        title: '일정이 삭제되었습니다.',
+        status: 'info',
+        duration: 3000,
+        isClosable: true,
+      });
+    } catch (error) {
+      console.error('Error deleting event:', error);
+      toast({
+        title: '일정 삭제 실패',
+        status: 'error',
+        duration: 3000,
+        isClosable: true,
+      });
+    }
+  };
+
+  const handleStartTimeChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const newStartTime = e.target.value;
+    setStartTime(newStartTime);
+    validateTime(newStartTime, endTime);
+  };
+
+  const handleEndTimeChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const newEndTime = e.target.value;
+    setEndTime(newEndTime);
+    validateTime(startTime, newEndTime);
   };
 
   const validateTime = (start: string, end: string) => {
@@ -314,12 +382,122 @@ function useEvents(): UseEventsReturn {
     }
   };
 
+  const checkUpcomingEvents = async () => {
+    const now = new Date();
+    const upcomingEvents = events.filter((event) => {
+      const eventStart = new Date(`${event.date}T${event.startTime}`);
+      const timeDiff = (eventStart.getTime() - now.getTime()) / (1000 * 60);
+      return (
+        timeDiff > 0 &&
+        timeDiff <= event.notificationTime &&
+        !notifiedEvents.includes(event.id)
+      );
+    });
+
+    for (const event of upcomingEvents) {
+      try {
+        setNotifications((prev) => [
+          ...prev,
+          {
+            id: event.id,
+            message: `${event.notificationTime}분 후 ${event.title} 일정이 시작됩니다.`,
+          },
+        ]);
+        setNotifiedEvents((prev) => [...prev, event.id]);
+      } catch (error) {
+        console.error('Error updating notification status:', error);
+      }
+    }
+  };
+
+  const getDaysInMonth = (year: number, month: number) => {
+    return new Date(year, month + 1, 0).getDate();
+  };
+
+  const getWeekDates = (date: Date) => {
+    const day = date.getDay();
+    const diff = date.getDate() - day + (day === 0 ? -6 : 1);
+    const monday = new Date(date.setDate(diff));
+    const weekDates = [];
+    for (let i = 0; i < 7; i++) {
+      const nextDate = new Date(monday);
+      nextDate.setDate(monday.getDate() + i);
+      weekDates.push(nextDate);
+    }
+    return weekDates;
+  };
+
+  const navigate = (direction: 'prev' | 'next') => {
+    setCurrentDate((prevDate) => {
+      const newDate = new Date(prevDate);
+      if (view === 'week') {
+        newDate.setDate(newDate.getDate() + (direction === 'next' ? 7 : -7));
+      } else if (view === 'month') {
+        newDate.setMonth(newDate.getMonth() + (direction === 'next' ? 1 : -1));
+      }
+      return newDate;
+    });
+  };
+
+  const searchEvents = (term: string) => {
+    if (!term.trim()) return events;
+
+    return events.filter(
+      (event) =>
+        event.title.toLowerCase().includes(term.toLowerCase()) ||
+        event.description.toLowerCase().includes(term.toLowerCase()) ||
+        event.location.toLowerCase().includes(term.toLowerCase())
+    );
+  };
+
+  const filteredEvents = (() => {
+    const filtered = searchEvents(searchTerm);
+    return filtered.filter((event) => {
+      const eventDate = new Date(event.date);
+      if (view === 'week') {
+        const weekDates = getWeekDates(currentDate);
+        return eventDate >= weekDates[0] && eventDate <= weekDates[6];
+      } else if (view === 'month') {
+        return (
+          eventDate.getMonth() === currentDate.getMonth() &&
+          eventDate.getFullYear() === currentDate.getFullYear()
+        );
+      }
+      return true;
+    });
+  })();
+
+  const formatWeek = (date: Date): string => {
+    const year = date.getFullYear();
+    const month = date.getMonth() + 1;
+    const weekNumber = Math.ceil(date.getDate() / 7);
+    return `${year}년 ${month}월 ${weekNumber}주`;
+  };
+
+  const formatMonth = (date: Date): string => {
+    const year = date.getFullYear();
+    const month = date.getMonth() + 1;
+    return `${year}년 ${month}월`;
+  };
+
   return {
     fetchHolidays,
     fetchEvents,
     addOrUpdateEvent,
     saveEvent,
+    editEvent,
+    deleteEvent,
     validateTime,
+    checkUpcomingEvents,
+    handleStartTimeChange,
+    handleEndTimeChange,
+    getDaysInMonth,
+    getWeekDates,
+    navigate,
+    searchEvents,
+    filteredEvents,
+    formatWeek,
+    formatMonth,
     events,
     setEvents,
     title,
